+++ migrations/2022-06-08-015851_create_items/up.sql
@@ -3,5 +3,5 @@ CREATE TABLE Item (
     item_id SERIAL PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
     description TEXT,
-    price DECIMAL(12, 2) NOT NULL
+    price FLOAT NOT NULL
 )
